Updated project before Meeting.
