+++ app/src/main/java/com/example/nyismaw/communitypolicing/ApiWrapper/DownloadFileInterface.java
@@ -1,0 +1,11 @@
+package com.example.nyismaw.communitypolicing.ApiWrapper;
+
+/**
+ * Created by nyismaw on 12/2/2017.
+ */
+
+public interface DownloadFileInterface {
+    public void downLoadImage(final String imageName);
+    public void downLoadAudio(final String audioName);
+
+}
