+++ app/src/main/java/com/example/nyismaw/communitypolicing/ApiWrapper/ImageBrowseInterface.java
@@ -1,0 +1,9 @@
+package com.example.nyismaw.communitypolicing.ApiWrapper;
+
+/**
+ * Created by nyismaw on 12/2/2017.
+ */
+
+public interface ImageBrowseInterface {
+    public void getImageByid(String id);
+}
