+++ app/src/main/java/com/example/nyismaw/communitypolicing/ApiWrapper/ReprotedIssuesInterface.java
@@ -1,0 +1,12 @@
+package com.example.nyismaw.communitypolicing.ApiWrapper;
+
+import java.util.List;
+
+/**
+ * Created by nyismaw on 12/2/2017.
+ */
+
+public interface ReprotedIssuesInterface  {
+    public void createObject(Object object, String description);
+    public List<Object> getReportedIssues();
+}
