+++ app/src/main/java/com/example/nyismaw/communitypolicing/ApiWrapper/UploadFileInterface.java
@@ -1,0 +1,12 @@
+package com.example.nyismaw.communitypolicing.ApiWrapper;
+
+import android.graphics.Bitmap;
+
+/**
+ * Created by nyismaw on 12/2/2017.
+ */
+
+public interface UploadFileInterface {
+    public void uploadAudio(String path, String fileName);
+    public void UploadImage(Bitmap bitmap, String filename);
+}
