+++ app/src/main/java/com/example/nyismaw/communitypolicing/AppInfo/CurrentLocation.java
@@ -1,4 +1,4 @@
-package com.example.nyismaw.communitypolicing.model;
+package com.example.nyismaw.communitypolicing.AppInfo;
 
 import android.location.Location;
