+++ app/src/main/java/com/example/nyismaw/communitypolicing/controller/filters/FilterIssues.java
@@ -1,4 +1,4 @@
-package com.example.nyismaw.communitypolicing.controller.filter;
+package com.example.nyismaw.communitypolicing.controller.filters;
 
 import com.example.nyismaw.communitypolicing.model.Issues;
 
@@ -11,6 +11,17 @@ import java.util.List;
 public class FilterIssues {
 
     public Issues filterIssueById(String id) {
+
+        List<Issues> issues = FetchedIssues.getIssues();
+        for (Issues iss : issues) {
+            if (iss.getId().equals(id))
+                return iss;
+        }
+
+        return null;
+    }
+
+    public Issues filterIssueByCategoryId(String id) {
 
         List<Issues> issues = FetchedIssues.getIssues();
         for (Issues iss : issues) {
