+++ app/src/main/java/com/example/nyismaw/communitypolicing/controller/login/SignInFactory.java
@@ -6,8 +6,8 @@ import com.example.nyismaw.communitypolicing.screens.SignInActivity;
  * Created by nyismaw on 11/28/2017.
  */
 
-public class LoginFactory {
-    public static LoginInterface getLogin(SignInActivity signInActivity, String loginType) {
+public class SignInFactory {
+    public static SignInInterface getLogin(SignInActivity signInActivity, String loginType) {
         if (loginType.equals("signInWithGoogle"))
             return new SignInWithGoogle(signInActivity);
         else if (loginType.equals("signInAnonumus"))
