+++ app/src/main/java/com/example/nyismaw/communitypolicing/controller/login/SignInInterface.java
@@ -4,7 +4,7 @@ package com.example.nyismaw.communitypolicing.controller.login;
  * Created by nyismaw on 11/28/2017.
  */
 
-public interface LoginInterface {
+public interface SignInInterface {
     public void signin();
     public void signout();
     public void authenticate(Object object);
