+++ app/src/main/java/com/example/nyismaw/communitypolicing/controller/notification/NotificationService.java
@@ -4,5 +4,5 @@ package com.example.nyismaw.communitypolicing.controller.notification;
  * Created by nyismaw on 11/28/2017.
  */
 
-public class CustomNotification {
+public class NotificationService {
 }
