+++ app/src/main/java/com/example/nyismaw/communitypolicing/exceptions/FIreBaseApiException.java
@@ -1,0 +1,8 @@
+package com.example.nyismaw.communitypolicing.exceptions;
+
+/**
+ * Created by nyismaw on 12/2/2017.
+ */
+
+public class FIreBaseApiException {
+}
